commit Attraction + boisson service et changement id en long

+++ src/main/java/model/Attraction.java
@@ -14,7 +14,7 @@ import javax.persistence.ManyToMany;
 public class Attraction {
 @Id
 @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 	private String nom;
 	private int duree, capacite, tailleMin, tailleMax;
 	private boolean restHandi;
@@ -26,7 +26,7 @@ public class Attraction {
 		
 	}
 	
-	public Attraction(int id, String nom, int duree, int capacite, int tailleMin, int tailleMax, boolean restHandi) {
+	public Attraction(Long id, String nom, int duree, int capacite, int tailleMin, int tailleMax, boolean restHandi) {
 		this.id=id;
 		this.nom=nom;
 		this.duree = duree;
@@ -45,10 +45,10 @@ public class Attraction {
 		this.restHandi = restHandi;
 	}
 	
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
